% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_idata.R
\name{as_idata}
\alias{as_idata}
\title{Create individual data frame from a covset object}
\usage{
as_idata(.covset, .n)
}
\arguments{
\item{.covset}{a covset object.}

\item{.n}{number of IDs to simulate.}
}
\value{
An idata set data frame.
}
\description{
Create individual data frame from a covset object
}
\examples{
cov1 <- covset(Y ~ rbinomial(0.2), Z ~ rnorm(2,2))

as_idata(cov1, 10)

}
