% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_covset.R, R/dmutate.R
\name{covset}
\alias{covset}
\alias{new_covobj}
\alias{print.covobj}
\alias{as.list,covobj-method}
\alias{as.list,covset-method}
\alias{print.covset}
\alias{rvset}
\alias{as.covset}
\title{Covobj and covset objects}
\usage{
new_covobj(x, envir = parent.frame(), ...)

\method{print}{covobj}(x, ...)

\S4method{as.list}{covobj}(x, ...)

\S4method{as.list}{covset}(x, ...)

\method{print}{covset}(x, ...)

covset(..., envir = parent.frame())

rvset(...)

as.covset(x)
}
\arguments{
\item{x}{a formula; may be quoted.}

\item{envir}{for formulae.}

\item{...}{formulae to use for the covset.}
}
\description{
Covobj and covset objects

Create a set of covariates
}
\details{
\code{rvset} is an alias for \code{covset}.
}
\examples{

obj <- new_covobj(Y[0,80] ~ rnorm(20,50))

obj

as.list(obj)

a <- Y ~ runif(0,1)
b <- Z ~ rbeta(1,1)

set <- covset(a,b)

set

as.list(set)

}
