% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cvec}
\alias{ch}
\alias{cvec}
\alias{s}
\title{Create create character vectors.}
\usage{
cvec(...)

ch(...)

s(...)
}
\arguments{
\item{...}{unquoted strings (for \code{ch})}

\item{x}{comma-separated quoted string (for \code{cvec})}
}
\description{
Create create character vectors.
}
\examples{

cvec("A,B,C")
ch(A,B,C)
s(A,B,C)

}

