% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Aaaa.R, R/dmutate.R
\name{dmutate}
\alias{dmutate}
\title{mutate a data frame, adding random variables.}
\usage{
dmutate(data, ...)
}
\arguments{
\item{data}{a data frame.}

\item{...}{formulae and other arguments for \code{\link[=mutate_random]{mutate_random()}}.}
}
\description{
mutate a data frame, adding random variables.

Apply formulae to a data frame
}
\examples{

idata <- dplyr::data_frame(ID = 1:10)

dmutate(idata, y ~ rbinomial(0.5), wt ~ rnorm(mu,sd),
        envir = list(mu = 50, sd = 20))

}
