% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmutate.R
\name{dmutate}
\alias{dmutate}
\title{Add random variates to your data frame.}
\usage{
dmutate(data, form, ...)
}
\arguments{
\item{data}{the data.frame to mutate}

\item{x}{an unquoted R formula; see details.}
}
\description{
Add random variates to your data frame.
}
\examples{


data(Theoph)

Theoph \%>\% dmutate(SEX~binom(0.3|ID)) \%>\% head
mu <- c(1,30)
Sigma <- diag(c(1,1))

Theoph \%>\% dmutate(CL+VC ~ MvLN(mu,Sigma|ID)) \%>\% head


}

