% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmutate.R
\name{rmvnorm}
\alias{rmvnorm}
\alias{rlmvnorm}
\alias{rmassnorm}
\alias{rlmassnorm}
\title{Simulate from multivariate normal distribution}
\usage{
rmvnorm(n, mu, Sigma)

rlmvnorm(n, ...)

rmassnorm(n, ...)

rlmassnorm(n, ...)
}
\arguments{
\item{n}{number of variates.}

\item{mu}{vector of means.}

\item{Sigma}{variance-covariance matrix with number of columns equal to
length of \code{mu}.}

\item{...}{arguments passed to \code{rmvnorm}.}
}
\value{
Returns a matrix of variates with number of rows
equal to \code{n} and number of columns equal to length of `mu.
}
\description{
Simulate from multivariate normal distribution
}
\details{
\code{rlmvnorm} is a multivariate log normal.

\code{rmassnorm} and \code{rlmassnorm} simulate the
multivariate normal using the \code{MASS} package.
}
