% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmutate.R
\name{rmvnorm}
\alias{rmvnorm}
\alias{rlmvnorm}
\title{Simulate from multivariate normal distribution.}
\usage{
rmvnorm(n, mu, Sigma)

rlmvnorm(n, ...)
}
\arguments{
\item{n}{number of variates}

\item{mu}{vector of means}

\item{Sigma}{variance-covariance matrix with number of columns equal to
length of \code{mu}}

\item{...}{arguments passed to \code{rmvnorm}}
}
\value{
Returns a matrix of variates with number of rows
equal to \code{n} and mumber of columns equal to length of \code{mu}.
}
\description{
Simulate from multivariate normal distribution.
}
\details{
\code{rlmvnorm} is a multivariate log normal.
}
